+++ src/lib.rs
@@ -93,6 +93,8 @@ pub use lzip::LZIPReader;
 pub use lzip::LZIPReaderMT;
 #[cfg(all(feature = "lzip", feature = "encoder"))]
 pub use lzip::{LZIPOptions, LZIPWriter};
+#[cfg(all(feature = "lzip", feature = "encoder", feature = "std"))]
+pub use lzip::LZIPWriterMT;
 pub use lzma2_reader::{get_memory_usage as lzma2_get_memory_usage, LZMA2Reader};
 #[cfg(feature = "std")]
 pub use lzma2_reader_mt::LZMA2ReaderMT;